update android tests
