+++ compose/material3/adaptive/adaptive-navigation3/src/androidInstrumentedTest/kotlin/androidx/compose/material3/adaptive/navigation3/ListDetailSceneStrategyScreenshotTest.kt
