+++ compose/material3/adaptive/adaptive-navigation3/src/androidInstrumentedTest/kotlin/androidx/compose/material3/adaptive/navigation3/ListDetailSceneStrategyTest.kt
@@ -20,29 +20,18 @@ package androidx.compose.material3.adaptive.navigation3
 
 import androidx.activity.OnBackPressedDispatcher
 import androidx.activity.compose.LocalOnBackPressedDispatcherOwner
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.size
 import androidx.compose.material3.adaptive.ExperimentalMaterial3AdaptiveApi
-import androidx.compose.material3.adaptive.currentWindowAdaptiveInfo
 import androidx.compose.material3.adaptive.layout.PaneScaffoldDirective
-import androidx.compose.material3.adaptive.layout.calculatePaneScaffoldDirective
 import androidx.compose.material3.adaptive.navigation.BackNavigationBehavior
-import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateListOf
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.assertIsNotDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithTag
-import androidx.compose.ui.unit.dp
+import androidx.compose.ui.test.onNodeWithText
 import androidx.navigation3.runtime.NavEntry
-import androidx.navigation3.runtime.entryProvider
 import androidx.navigation3.scene.Scene
-import androidx.navigation3.ui.NavDisplay
+import androidx.navigation3.scene.SceneStrategyScope
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.MediumTest
 import com.google.common.truth.Truth.assertThat
@@ -56,12 +45,6 @@ import org.junit.runner.RunWith
 class ListDetailSceneStrategyTest {
     @get:Rule val composeRule = createComposeRule(StandardTestDispatcher())
 
-    val HomeScreenTestTag = "HomeScreen"
-    val ListScreenTestTag = "ListScreen"
-    val DetailScreenTestTag = "DetailScreen"
-    val DetailPlaceholderScreenTestTag = "PlaceholderScreen"
-    val ExtraScreenTestTag = "ExtraScreen"
-
     @Test
     fun calculateScene_singlePane_nullListDetailScene() {
         var scene: Scene<TestKey>? = null
@@ -69,7 +52,7 @@ class ListDetailSceneStrategyTest {
         composeRule.setContent {
             val strategy =
                 rememberListDetailSceneStrategy<TestKey>(directive = PaneScaffoldDirective.Default)
-            scene = strategy.calculateScene(listOf(nonListDetailEntry, listEntry, detailEntry)) {}
+            scene = strategy.calculateScene(listOf(nonListDetailEntry, listEntry, detailEntry))
         }
 
         composeRule.waitForIdle()
@@ -83,7 +66,7 @@ class ListDetailSceneStrategyTest {
         composeRule.setContent {
             val strategy =
                 rememberListDetailSceneStrategy<TestKey>(directive = MockDualPaneScaffoldDirective)
-            scene = strategy.calculateScene(entries) {}
+            scene = strategy.calculateScene(entries)
         }
 
         composeRule.waitForIdle()
@@ -96,7 +79,7 @@ class ListDetailSceneStrategyTest {
         composeRule.setContent {
             val strategy =
                 rememberListDetailSceneStrategy<TestKey>(directive = MockDualPaneScaffoldDirective)
-            scene = strategy.calculateScene(listOf(nonListDetailEntry, listEntry, detailEntry)) {}
+            scene = strategy.calculateScene(listOf(nonListDetailEntry, listEntry, detailEntry))
         }
 
         composeRule.waitForIdle()
@@ -109,7 +92,7 @@ class ListDetailSceneStrategyTest {
         composeRule.setContent {
             val strategy =
                 rememberListDetailSceneStrategy<TestKey>(directive = MockDualPaneScaffoldDirective)
-            scene = strategy.calculateScene(listOf(listEntry, detailEntry, nonListDetailEntry)) {}
+            scene = strategy.calculateScene(listOf(listEntry, detailEntry, nonListDetailEntry))
         }
 
         composeRule.waitForIdle()
@@ -122,7 +105,7 @@ class ListDetailSceneStrategyTest {
         composeRule.setContent {
             val strategy =
                 rememberListDetailSceneStrategy<TestKey>(directive = MockDualPaneScaffoldDirective)
-            scene = strategy.calculateScene(listOf(nonListDetailEntry, listEntry)) {}
+            scene = strategy.calculateScene(listOf(nonListDetailEntry, listEntry))
         }
 
         composeRule.waitForIdle()
@@ -135,7 +118,7 @@ class ListDetailSceneStrategyTest {
         composeRule.setContent {
             val strategy =
                 rememberListDetailSceneStrategy<TestKey>(directive = MockDualPaneScaffoldDirective)
-            scene = strategy.calculateScene(listOf(nonListDetailEntry, detailEntry)) {}
+            scene = strategy.calculateScene(listOf(nonListDetailEntry, detailEntry))
         }
 
         composeRule.waitForIdle()
@@ -148,8 +131,7 @@ class ListDetailSceneStrategyTest {
         composeRule.setContent {
             val strategy =
                 rememberListDetailSceneStrategy<TestKey>(directive = MockDualPaneScaffoldDirective)
-            scene =
-                strategy.calculateScene(listOf(listEntry, detailEntry, listEntry, detailEntry)) {}
+            scene = strategy.calculateScene(listOf(listEntry, detailEntry, listEntry, detailEntry))
         }
 
         composeRule.waitForIdle()
@@ -162,7 +144,7 @@ class ListDetailSceneStrategyTest {
         composeRule.setContent {
             val strategy =
                 rememberListDetailSceneStrategy<TestKey>(directive = MockDualPaneScaffoldDirective)
-            scene = strategy.calculateScene(listOf(listEntry, nonListDetailEntry, detailEntry)) {}
+            scene = strategy.calculateScene(listOf(listEntry, nonListDetailEntry, detailEntry))
         }
 
         composeRule.waitForIdle()
@@ -178,7 +160,7 @@ class ListDetailSceneStrategyTest {
             scene =
                 strategy.calculateScene(
                     listOf(listEntry, detailEntry, nonListDetailEntry, listEntry, detailEntry)
-                ) {}
+                )
         }
 
         composeRule.waitForIdle()
@@ -194,7 +176,7 @@ class ListDetailSceneStrategyTest {
             scene =
                 strategy.calculateScene(
                     listOf(nonListDetailEntry, listEntry, detailEntry, detailEntry)
-                ) {}
+                )
         }
 
         composeRule.waitForIdle()
@@ -210,7 +192,7 @@ class ListDetailSceneStrategyTest {
             scene =
                 strategy.calculateScene(
                     listOf(nonListDetailEntry, listEntry, listEntry, detailEntry)
-                ) {}
+                )
         }
 
         composeRule.waitForIdle()
@@ -229,7 +211,7 @@ class ListDetailSceneStrategyTest {
         composeRule.setContent {
             val strategy =
                 rememberListDetailSceneStrategy<TestKey>(directive = MockDualPaneScaffoldDirective)
-            scene = strategy.calculateScene(listOf(listEntry1, listEntry2)) {}
+            scene = strategy.calculateScene(listOf(listEntry1, listEntry2))
         }
 
         composeRule.waitForIdle()
@@ -277,6 +259,25 @@ class ListDetailSceneStrategyTest {
     }
 
     @Test
+    fun dualPane_backstackWithListDetail_navigate_showsNewDetail() {
+        val backStack = mutableStateListOf(HomeKey, ListKey, DetailKey("abc"))
+        composeRule.setContent {
+            NavScreen(backStack = backStack, directive = MockDualPaneScaffoldDirective)
+        }
+        composeRule.onNodeWithTag(ListScreenTestTag).assertIsDisplayed()
+        composeRule.onNodeWithTag(DetailScreenTestTag).assertIsDisplayed()
+        composeRule.onNodeWithText("Detail(abc)").assertIsDisplayed()
+
+        backStack.add(DetailKey("def"))
+
+        composeRule.waitForIdle()
+        composeRule.onNodeWithTag(ListScreenTestTag).assertIsDisplayed()
+        composeRule.onNodeWithTag(DetailScreenTestTag).assertIsDisplayed()
+        composeRule.onNodeWithText("Detail(abc)").assertIsNotDisplayed()
+        composeRule.onNodeWithText("Detail(def)").assertIsDisplayed()
+    }
+
+    @Test
     fun dualPane_backstackWithListDetailExtra_showsDetailAndExtra() {
         val backStack = mutableStateListOf(HomeKey, ListKey, DetailKey("abc"), ExtraKey("abc"))
         composeRule.setContent {
@@ -296,9 +297,15 @@ class ListDetailSceneStrategyTest {
             backPressedDispatcher =
                 LocalOnBackPressedDispatcherOwner.current!!.onBackPressedDispatcher
         }
+        composeRule.onNodeWithTag(DetailScreenTestTag).assertIsDisplayed()
+        composeRule.onNodeWithTag(ExtraScreenTestTag).assertIsDisplayed()
+
         composeRule.runOnIdle { backPressedDispatcher.onBackPressed() }
 
         assertThat(backStack).containsExactly(HomeKey, ListKey, DetailKey("abc")).inOrder()
+        composeRule.onNodeWithTag(DetailScreenTestTag).assertIsDisplayed()
+        composeRule.onNodeWithTag(ListScreenTestTag).assertIsDisplayed()
+        composeRule.onNodeWithTag(ExtraScreenTestTag).assertIsNotDisplayed()
     }
 
     @Test
@@ -314,9 +321,16 @@ class ListDetailSceneStrategyTest {
             backPressedDispatcher =
                 LocalOnBackPressedDispatcherOwner.current!!.onBackPressedDispatcher
         }
+        composeRule.onNodeWithTag(DetailScreenTestTag).assertIsDisplayed()
+        composeRule.onNodeWithTag(ListScreenTestTag).assertIsDisplayed()
+
         composeRule.runOnIdle { backPressedDispatcher.onBackPressed() }
 
+        composeRule.waitForIdle()
+
         assertThat(backStack).containsExactly(HomeKey, ListKey).inOrder()
+        composeRule.onNodeWithTag(DetailScreenTestTag).assertIsNotDisplayed()
+        composeRule.onNodeWithTag(ListScreenTestTag).assertIsDisplayed()
     }
 
     @Test
@@ -328,64 +342,14 @@ class ListDetailSceneStrategyTest {
             backPressedDispatcher =
                 LocalOnBackPressedDispatcherOwner.current!!.onBackPressedDispatcher
         }
+        composeRule.onNodeWithTag(DetailScreenTestTag).assertIsDisplayed()
+        composeRule.onNodeWithTag(ListScreenTestTag).assertIsDisplayed()
+
         composeRule.runOnIdle { backPressedDispatcher.onBackPressed() }
 
+        composeRule.onNodeWithTag(DetailScreenTestTag).assertIsNotDisplayed()
+        composeRule.onNodeWithTag(ListScreenTestTag).assertIsNotDisplayed()
         assertThat(backStack).containsExactly(HomeKey).inOrder()
-    }
-
-    @Composable
-    fun NavScreen(
-        backStack: List<TestKey>,
-        backNavigationBehavior: BackNavigationBehavior =
-            BackNavigationBehavior.PopUntilScaffoldValueChange,
-        directive: PaneScaffoldDirective =
-            calculatePaneScaffoldDirective(currentWindowAdaptiveInfo()),
-    ) {
-        val listDetailSceneStrategy =
-            rememberListDetailSceneStrategy<TestKey>(
-                backNavigationBehavior = backNavigationBehavior,
-                directive = directive,
-            )
-        NavDisplay(
-            backStack = backStack,
-            modifier = Modifier.fillMaxSize(),
-            sceneStrategy = listDetailSceneStrategy,
-            entryProvider =
-                entryProvider {
-                    entry<HomeKey> {
-                        Box(Modifier.testTag(HomeScreenTestTag).size(100.dp).background(Color.Red))
-                    }
-
-                    entry<ListKey>(
-                        metadata =
-                            ListDetailSceneStrategy.listPane(ListKey) {
-                                Box(
-                                    Modifier.testTag(DetailPlaceholderScreenTestTag)
-                                        .size(100.dp)
-                                        .background(Color.Cyan)
-                                )
-                            }
-                    ) {
-                        Box(
-                            Modifier.testTag(ListScreenTestTag).size(100.dp).background(Color.Green)
-                        )
-                    }
-
-                    entry<DetailKey>(metadata = ListDetailSceneStrategy.detailPane(ListKey)) {
-                        Box(
-                            Modifier.testTag(DetailScreenTestTag)
-                                .size(100.dp)
-                                .background(Color.Blue)
-                        )
-                    }
-
-                    entry<ExtraKey>(metadata = ListDetailSceneStrategy.extraPane(ListKey)) {
-                        Box(
-                            Modifier.testTag(ExtraScreenTestTag).size(100.dp).background(Color.Gray)
-                        )
-                    }
-                },
-        )
     }
 }
 
@@ -397,8 +361,5 @@ private val detailEntry: NavEntry<TestKey> =
 
 private val nonListDetailEntry: NavEntry<TestKey> = NavEntry(HomeKey) {}
 
-private val MockDualPaneScaffoldDirective =
-    PaneScaffoldDirective.Default.copy(
-        maxHorizontalPartitions = 2,
-        horizontalPartitionSpacerSize = 16.dp,
-    )
+private fun ListDetailSceneStrategy<TestKey>.calculateScene(entries: List<NavEntry<TestKey>>) =
+    with(this) { SceneStrategyScope<TestKey>().calculateScene(entries) }
